requirements.txt
Jinja2==3.1.6
openpyxl==3.1.5
pandas==2.2.3
python_dotenv-1.1.0